+++ CHANGELOG.md
@@ -11,7 +11,9 @@ Note that this project **does not** adhere to [Semantic Versioning](https://semv
 
 ### Added
 
-- We added `POST /libraries/query` to the HTTP server to search for entries by DOI or URL across all open libraries in a single request. [#15599](https://github.com/JabRef/jabref/pull/15599)
+- We added `POST /libraries/query` to the HTTP server to search for entries across all open libraries with a list of search queries in a single request. [#15599](https://github.com/JabRef/jabref/pull/15599)
+- We added `GET /libraries/{id}/groups` to the HTTP server to list a library's groups as a flat breadcrumb list. [#15599](https://github.com/JabRef/jabref/pull/15599)
+- We added an optional `group` query parameter to `POST /libraries/{id}/entries` so imported entries can be assigned to a group, which is created if it does not exist. [#15599](https://github.com/JabRef/jabref/pull/15599)
 
 ### Changed